HW#5 Moved dao package from service to domain
